drivers: adc-dac: ad74416h: fix type mismatch in get_adc_single
ad74416h_get_adc_single() takes a uint16_t *val parameter but passes
it directly to ad74416h_get_raw_adc_result() which expects uint32_t *.
On AD74416H devices, get_raw_adc_result writes a 24-bit value through
a 32-bit pointer, causing a 4-byte write to a 2-byte variable and
corrupting adjacent stack memory.

Change the function prototype to uint32_t * to match the actual ADC
result width and align with get_raw_adc_result.

Fixes: 638bc3f8ba8d ("drivers:adc-dac:ad74416h: initial commit")

+++ drivers/adc-dac/ad74416h/ad74416h.h
@@ -718,7 +718,7 @@ int ad74416h_set_adc_conv_mux(struct ad74416h_desc *, uint32_t,
 int ad74416h_set_adc_conv_seq(struct ad74416h_desc *, enum ad74416h_conv_seq);
 
 /** Get a single ADC raw value for a specific channel, then power down the ADC */
-int ad74416h_get_adc_single(struct ad74416h_desc *, uint32_t, uint16_t *);
+int ad74416h_get_adc_single(struct ad74416h_desc *, uint32_t, uint32_t *);
 
 /** Read the die's temperature from the diagnostic register */
 int ad74416h_get_temp(struct ad74416h_desc *, uint32_t, uint16_t *);
